+++ packages/molecule/src/bytes.ts
@@ -1,0 +1,10 @@
+/**
+ * Codec for bytes
+ * @module
+ */
+import { byteArray, byteFixvec } from "./core";
+
+export const Byte10 = byteArray("Byte10", 10);
+export const Byte20 = byteArray("Byte10", 20);
+export const Byte32 = byteArray("Byte32", 32);
+export const Bytes = byteFixvec("Bytes");