+++ packages/molecule/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./externals";
-import * as mol from "./externals";
 
+import * as mol from "./externals";
 export { mol };
 export default mol;